requirements.txt
# Core RAG and ML dependencies
streamlit==1.28.1
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Vector database and embeddings
faiss-cpu==1.7.4
pinecone-client==2.2.4
weaviate-client==3.25.3

# Data processing and visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2

# Web framework and UI
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
streamlit-chat==0.1.1

# File handling and utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
python-docx==1.1.0

# Database and storage

pickle-mixin==1.0.2

# Evaluation and metrics
ragas==0.0.22
evaluate==0.4.1
rouge-score==0.1.2

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Additional utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0 